requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai>=2.7.0
requests==2.31.0
pdfplumber==0.10.4
